Archive unifyOffIceLabels migration
