+++ eisbuk-admin/functions/src/archive/oldMigrations.ts
@@ -12,10 +12,12 @@ import {
   Category,
   CustomerBase,
   OrgSubCollection,
+  SlotType,
 } from "eisbuk-shared";
 import {
   DeprecatedOrgSubCollection,
   DeprecatedBookingsMeta,
+  DeprecatedSlotType,
 } from "eisbuk-shared/dist/deprecated";
 
 import { __functionsZone__ } from "../constants";
@@ -311,4 +313,36 @@ export const addIdsToCustomers = functions
       return { success: false };
     }
   });
+
+/**
+ * Replaces all deprecated "off-ice-dancing" and "off-ice-gym" slot type entries
+ * with unified "off-ice" type
+ */
+export const unifyOffIceLabels = functions
+  .region("europe-west6")
+  .https.onCall(async ({ organization }, { auth }) => {
+    checkUser(organization, auth);
+
+    const firestore = admin.firestore();
+    const batch = firestore.batch();
+    // get all off-ice-* slots
+    const slotsRef = firestore
+      .collection(Collection.Organizations)
+      .doc(organization)
+      .collection(OrgSubCollection.Slots);
+    const offIceSlots = await slotsRef
+      .where("type", "in", [
+        DeprecatedSlotType.OffIceDancing,
+        DeprecatedSlotType.OffIceGym,
+      ])
+      .get();
+    // replace all type entries with unified "off-ice"
+    offIceSlots.forEach((snap) => {
+      const { ref } = snap;
+      batch.set(ref, { type: SlotType.OffIce }, { merge: true });
+    });
+    await batch.commit();
+    return { success: true };
+  });
+
 // #endregion addIdsToCustomers
